Refactoring API for GameFormat and Collection (sorting, toString, +NewMethods)
Most of the methods to return lists of formats now are "cached" since format lists is initialised at loading-time and never changed during the game.
This is particularly relevant for Archived formats which are a lot!.
To ease format organisation in filter menu, the API now includes a new method that organises archived formats per Subtype and Name-Prefix.

The natural order is always kept and taken into account, for each grouping.

On this note, the `compareTo` method has been revised so that there is always a fallback on name comparison. However, this is the last resort for Archived formats which do consider the EffectiveDate. When not available (e.g. Block formats in "formats/Blocks" folder), the index is used before the name.

Last but not least, GameFormat.toString now includes the Effective date in its name to better identify the format in the list.
(Particular useful with filters)